refactor: address PR feedback on identity command refactor
