+++ package.json
@@ -25,5 +25,4 @@
     "typescript": "^4.9.5"
   },
   "version": "0.0.2-beta.0",
-  "packageManager": "pnpm@9.15.4+sha512.b2dc20e2fc72b3e18848459b37359a32064663e5627a51e4c74b2c29dd8e8e0491483c3abb40789cfd578bf362fb6ba8261b05f0387d76792ed6e23ea3b1b6a0"
 }
